Move CoordinatorServiceClient to druid-server for use by extensions
